+++ core/consensus/src/namespaced_pipeline.rs
@@ -324,7 +324,7 @@ mod tests {
         Message::<PrepareHeader>::new(std::mem::size_of::<PrepareHeader>()).transmute_header(
             |_, new| {
                 *new = PrepareHeader {
-                    command: Command2::Prepare as u8,
+                    command: Command2::Prepare,
                     op,
                     parent,
                     checksum,
